+++ NexmoClient/Internal/NexmoCore/Network/SocketClient/NXMSocketClient.m
@@ -240,7 +240,6 @@
 
 - (void)didFailLoginWithError:(NXMErrorCode)error {
     self.token = nil;
-    [self disconnectSocket];
     
     NXMConnectionStatusReason reason = NXMConnectionStatusReasonUnknown;
     
@@ -255,11 +254,16 @@
         case NXMErrorCodeTokenExpired:
             reason = NXMConnectionStatusReasonTokenExpired;
             break;
+        case NXMErrorCodeUserNotFound:
+            reason = NXMConnectionStatusReasonUserNotFound;
+            break;
         default:
             break;
     }
     
     [self updateConnetionStatus:NXMConnectionStatusDisconnected reason:reason];
+    
+    [self disconnectSocket];
 }
 
 - (void)didServerLoginWithData:(NSArray *)data {
@@ -315,6 +319,8 @@
 }
 
 - (void)subscribeGeneralEvents {
+    __weak NXMSocketClient *weakSelf = self;
+    
     [self.socket on:kNXMSocketEventBadPermission callback:^(NSString *event, NSArray *data, VPSocketAckEmitter *emitter) {
         LOG_ERROR("socket BadPermission");
     }];
@@ -325,7 +331,7 @@
     
     [self.socket on:kNXMSocketEventUserNotFound callback:^(NSString *event, NSArray *data, VPSocketAckEmitter *emitter) {
         LOG_ERROR("socket kNXMSocketEventUserNotFound"  );
-        //TODO: check if this means anything about login/logout and also regard the invaliduser sessioninternalerror
+        [weakSelf didFailLoginWithError:NXMErrorCodeUserNotFound];
     }];
 }